+++ default.nix
@@ -1,6 +1,6 @@
 { stdenv, lib, jq, nix, makeWrapper }:
 
-stdenv.mkDerivation rec {
+stdenv.mkDerivation {
   pname = "terranix";
   version = "2.9.0";
 
@@ -15,12 +15,4 @@ stdenv.mkDerivation rec {
     wrapProgram $out/bin/terranix-doc-json \
       --prefix PATH : ${lib.makeBinPath [ jq nix ]}
   '';
-
-  meta = with lib; {
-    description = "A NixOS like terraform-json generator";
-    homepage = "https://terranix.org";
-    license = licenses.gpl3;
-    platforms = platforms.unix;
-    maintainers = with maintainers; [ mrVanDalo ];
-  };
 }